update LightGBMTabularTrees to inherit from BaseModelTabularTrees
